+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] -
+### Changed
+- Access to story.virtualAnimator is deprecated.
+  > use addVirtualAnimator/removeVirtualAnimator instead.
+
+### Added
+- Add addVirtualAnimator/removeVirtualAnimator to the story.
+
+
 ## [1.0.1] -
 ### Fix
 - Update cdn.rawgit.com links with a-bentofreire.github.io.